Clean up rough edges
