+++ swagger/types.go
@@ -1,0 +1,49 @@
+package swagger
+
+import (
+	"fmt"
+	"strings"
+
+	"github.com/go-openapi/spec"
+)
+
+// TypeFromSchema returns the type of a Swagger schema as a string. If includeModels is true
+// then it returns models.TYPE
+func TypeFromSchema(schema *spec.Schema, includeModels bool) (string, error) {
+	// We support three types of schemas
+	// 1. An empty schema, which we represent by the empty struct
+	// 2. A schema with one element, the $ref key
+	// 3. A schema with two elements. One a type with value 'array' and another items field
+	// referencing the $ref
+	if schema == nil {
+		return "struct{}", nil
+	} else if schema.Ref.String() != "" {
+		ref := schema.Ref.String()
+		if !strings.HasPrefix(ref, "#/definitions/") {
+			return "", fmt.Errorf("schema.$ref has undefined reference type. Must start with #/definitions")
+		}
+		def := ref[len("#/definitions/"):]
+		if includeModels {
+			def = "models." + def
+		}
+		return def, nil
+	} else {
+		schemaType := schema.Type
+		if len(schemaType) != 1 || schemaType[0] != "array" {
+			return "", fmt.Errorf("Two element schemas must have a 'type' field with the value 'array'")
+		}
+		items := schema.Items
+		if items == nil || items.Schema == nil {
+			return "", fmt.Errorf("Two element schemas must have a '$ref' field in the 'items' descriptions")
+		}
+		ref := items.Schema.Ref.String()
+		if !strings.HasPrefix(ref, "#/definitions/") {
+			return "", fmt.Errorf("schema.$ref has undefined reference type. Must start with #/definitions")
+		}
+		def := ref[len("#/definitions/"):]
+		if includeModels {
+			def = "models." + def
+		}
+		return "[]" + def, nil
+	}
+}
